fabtests: Create RPC lib
Move RPC related code to common file and define corresponding header.
